+++ backend/README.md
@@ -1,0 +1,11 @@
+# Backend
+
+## Running Tests
+
+Install dependencies and run:
+
+```bash
+npm test
+```
+
+This executes Jest with Supertest to test the API endpoints.
